+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make `includes` use glob patterns by @daimond113
 - Use symlinks for workspace dependencies to not require reinstalling by @daimond113
 - Add `auth token` command to print the auth token for the index by @daimond113
+- Support specifying which external registries are allowed on registries by @daimond113
 
 ### Fixed
 - Install dependencies of packages in `x` command